# openapi.yaml
openapi: 3.0.3
info:
  title: Bloom Health API
  version: 1.0.1
  description: REST API for Bloom Health therapy practice platform
  contact:
    name: Bloom Health
    url: https://bloomhealth.us
servers:
  - url: http://localhost:3000/api
    description: Local development
  - url: https://bloomhealth.us/api
    description: Production
  - url: https://qa.bloomhealth.us/api
    description: QA/Staging
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: better-auth.session_token
      description: Session cookie set by Better Auth
  schemas:
...
    Conversation:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        avatar:
          type: string
        avatarColor:
          type: string
        lastMessage:
          type: string
        time:
          type: string
        unread:
          type: integer
        active:
          type: boolean
    Message:
      type: object
      properties:
        id:
          type: string
        sender:
          type: string
        message:
          type: string
        time:
          type: string
        isMe:
          type: boolean
        avatar:
          type: string
        avatarColor:
          type: string
        reactions:
          type: array
          items:
            type: object
            properties:
              emoji:
                type: string
              count:
                type: integer
...
    UploadResponse:
      type: object
      properties:
        url:
          type: string
        pathname:
          type: string
...
    Error:
      type: object
      properties:
        error:
          type: string
    Success:
      type: object
      properties:
        success:
          type: boolean
security:
  - cookieAuth: []
paths:
...
  /messages:
    get:
      tags: [Messages]
      summary: Get conversations and messages
      parameters:
        - name: conversationId
          in: query
          schema:
            type: string
          description: Optional conversation id to focus
      responses:
        '200':
          description: Conversations and messages
          content:
            application/json:
              schema:
                type: object
                properties:
                  conversations:
                    type: array
                    items:
                      $ref: '#/components/schemas/Conversation'
                  messages:
                    type: array
                    items:
                      $ref: '#/components/schemas/Message'
        '401':
          description: Unauthorized
    post:
      tags: [Messages]
      summary: Send a message or start a conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                conversationId:
                  type: string
                  description: Existing conversation id
                recipientId:
                  type: string
                  description: Target user id (used when no conversation exists)
                message:
                  type: string
                  description: Message content (required unless startOnly = true)
                startOnly:
                  type: boolean
                  description: When true, ensure conversation exists without sending a message
      responses:
        '200':
          description: Message sent or conversation created
          content:
            application/json:
              schema:
                type: object
                properties:
                  conversation:
                    $ref: '#/components/schemas/Conversation'
                  message:
                    $ref: '#/components/schemas/Message'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not a participant in the conversation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Recipient or conversation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /upload:
    post:
      tags: [Files]
      summary: Upload a file to blob storage
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Upload success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
        '400':
          description: No file provided
        '401':
          description: Unauthorized
        '500':
          description: Upload misconfigured or failed
...
  /user/password:
    post:
      tags: [User]
      summary: Change password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword: { type: string }
                newPassword: { type: string, minLength: 8 }
      responses:
        '200':
          description: Password changed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '500':
          description: Failed to change password
...